requirements.txt
streamlit>=1.36.0
google-generativeai>=0.5.0